navegacao estruturada
